feat: add limits checking
this checks if the number specified fits within the maximum int size

+++ src/header.h
@@ -3,6 +3,7 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <limits.h>
 #include <string.h>
 #include <time.h>
 
@@ -14,7 +15,6 @@
 	extern const char *program_name;
 	extern const bool verbose;
 
-
 void clear(void);
 void err(const char *error_message);
 int command_parsing(int num_args, char *arguments[]);
@@ -23,6 +23,6 @@ int time_parsing(int *flag_r_index, int num_args, char *argument[]);
 void stotime(int total_seconds, int *hours, int *minutes, int *seconds);
 
 void date_input(void);
-long get_time(int lower_bound, int upper_bound, char *date_type);
+int get_time(int lower_bound, int upper_bound, char *date_type);
 
 void wait(int quarters_seconds, int num_seconds);
